backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
groq
python-dotenv
pydantic
firebase-admin
google-generativeai